Killed warning

+++ device.c
@@ -1144,7 +1144,7 @@ int cXinelibDevice::TsBufferFlush(void)
 {
   if (m_TsBufSize) {
     int n;
-    if ((n = PlayAny(m_TsBuf, m_TsBufSize)) == m_TsBufSize) {
+    if ((n = PlayAny(m_TsBuf, m_TsBufSize)) == (int)m_TsBufSize) {
       m_TsBufSize = 0;
       return n;
     }
@@ -1193,7 +1193,6 @@ int cXinelibDevice::PlayTsAudio(const uchar *Data, int Length)
 
 int cXinelibDevice::PlayTsVideo(const uchar *Data, int Length)
 {
-#warning PlayVideo: get_size, trickspeed stuff, ... !!!
   return PlayTsAny(Data, Length);
 }
 #endif // VDRVERSNUM >= 10701
